+++ eslint.config.js
@@ -1,12 +1,23 @@
 import tseslint from "typescript-eslint";
+import { defineConfig } from "eslint/config";
 import eslintPlugin from "eslint-plugin-eslint-plugin";
 import node from "eslint-config-mlauffer-nodejs";
 
-export default tseslint.config([
+export default defineConfig([
   {
     ignores: ["dist/", "coverage/", "docs/", "test/"],
   },
   {
+    plugins: {
+      "@typescript-eslint": tseslint.plugin,
+    },
+    languageOptions: {
+      parser: tseslint.parser,
+      parserOptions: {
+        projectService: true,
+        tsconfigRootDir: import.meta.dirname,
+      },
+    },
     extends: [node, eslintPlugin.configs.recommended, tseslint.configs.strict],
     rules: {
       "jsdoc/require-param": "off",
@@ -15,4 +26,8 @@ export default tseslint.config([
       "sonarjs/no-skipped-tests": "warn",
     },
   },
+  {
+    files: ['**/*.js'],
+    extends: [tseslint.configs.disableTypeChecked],
+  }
 ]);